refactor(ui): move card grouping into DocumentationPage
Reworks the kind-based card grouping so the logic lives in the page
renderer rather than as a `grouped` prop on the shared `TreeCards`.

`DocumentationPage` now pre-filters its children per `kind` with
`queryElements({ "props.kind": … })` and renders each non-empty group
as its own `<Section>` + `<Heading>` + `<TreeCards>`. `TreeCards`,
`FilePage`, and the tree README revert to their original form.

The docs site maps `tree-file` pages to `DocumentationPage` too (via
`TreePageMapping`) — a file's props are a compatible subset of a
symbol's — so module/file pages get the same grouped sections.
